feat: add shared utilities module
